requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
alembic==1.13.2
httpx==0.27.0

# Vector similarity search
pgvector==0.4.2

# Blob storage
boto3==1.35.0

# Security
passlib[bcrypt]==1.7.4

# Observability
prometheus-fastapi-instrumentator==7.1.0
structlog==25.5.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
